requirements.txt
python-dotenv
requests
ollama
groq
streamlit
openai
sentence-transformers 